% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allow_named_choices}
\alias{allow_named_choices}
\title{Allow named choices}
\usage{
allow_named_choices(
  inputId,
  update_function,
  input,
  session,
  init_choices,
  init_selected = NULL,
  ...
)
}
\arguments{
\item{inputId}{id of input widget}

\item{update_function}{to handle updates to widget}

\item{input}{def}

\item{session}{def}

\item{init_choices}{named choices to initialize extended widget}

\item{init_selected}{named select  to initialize extended widget}

\item{...}{ellipsis innit?}
}
\value{
a list of functions:
read() reads a named input from widget
update() update widget with named choices og selection (by name)
}
\description{
Allow named choices
}
